Add files via upload
Posting current progress. There is an error in the response from the controller when I do an http.request to Github and I don't know what's jamming it up. On the off hand chance that it's something local on my dev box, I am going to attempt a docker deployment on a different machine.